requirements.txt
Flask
Flask-Cors
Flask-SocketIO
Flask-SQLAlchemy
Werkzeug
pandas
numpy
joblib
requests
twilio
python-dotenv
email-validator
eventlet
